refactor(course-form): extract business logic into useCourseForm hook
CourseForm was violating SRP by handling form state, validation, author
management, Redux dispatching, and navigation all in one component.

- Created src/components/CourseForm/hooks/useCourseForm.js containing
  all form logic: state, validation, handlers, dispatch, navigation
- CourseForm component is now a pure rendering layer with no direct
  Redux or navigation dependencies
- Hook returns a structured object split into fields, setters, state
  and handlers groups for clear intent communication
- validateCourseFields and validation constants moved into the hook
  file as private implementation details
